+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+## [0.1.4] - 2024-07-03
+### Added
+- `AmrDocument` infers its document ID from the first contained sentence.
+- Added feature sentence and documents from AMR analogues in
+  `AnnotatedAmrFeatureDocumentFactory`.
+
+
 ## [0.1.3] - 2024-05-16
 ### Added
 - "A utility class to reindex variables in an `AmrDocument` and a decorator for
@@ -64,7 +71,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 <!-- links -->
-[Unreleased]: https://github.com/Paul Landes/amr/compare/v0.1.3...HEAD
+[Unreleased]: https://github.com/Paul Landes/amr/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/Paul Landes/amr/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/Paul Landes/amr/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/Paul Landes/amr/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/Paul Landes/amr/compare/v0.1.0...v0.1.1